@@ -1,12 +1,15 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
+    
+    <Nullable>enable</Nullable><ImplicitUsings>enable</ImplicitUsings>
+    
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="DevExpress.AspNetCore.Dashboard" Version="21.2.3" />
-    <PackageReference Include="DevExpress.AspNetCore.Dashboard.de" Version="21.2.3" />
-    <PackageReference Include="DevExpress.Blazor.Dashboard" Version="21.2.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Localization" Version="2.2.0" />
+    <PackageReference Include="DevExpress.AspNetCore.Dashboard" Version="22.1.3" />
+    <PackageReference Include="DevExpress.AspNetCore.Dashboard.de" Version="22.1.3" />
+    <PackageReference Include="DevExpress.Blazor.Dashboard" Version="22.1.3" />
+    <PackageReference Include="System.Data.SQLite.Core" Version="1.0.115.5" />
   </ItemGroup>
-
-  </Project>
+</Project>
