@@ -1,10 +1,10 @@
-﻿using DevExpress.DashboardAspNetCore;
+﻿﻿﻿using DevExpress.DashboardAspNetCore;
 using DevExpress.DashboardWeb;
 using Microsoft.AspNetCore.DataProtection;
 
-namespace BlazorDashboardApp {
+namespace BlazorDashboardApp.Controllers {
     public class DefaultDashboardController : DashboardController {
-        public DefaultDashboardController(DashboardConfigurator configurator, IDataProtectionProvider dataProtectionProvider = null)
+        public DefaultDashboardController(DashboardConfigurator configurator, IDataProtectionProvider? dataProtectionProvider = null)
             : base(configurator, dataProtectionProvider) {
         }
     }
