@@ -27,9 +27,9 @@ namespace BlazorDashboardApp {
         }
 
         public int ID { get; set; }
-        public string Name { get; set; }
+        public string ?Name { get; set; }
         public DateTime HireDate { get; set; }
-        public string Department { get; set; }
+        public string ?Department { get; set; }
         public double Salary { get; set; }
 
     }
