@@ -6,7 +6,7 @@
 
 <CultureSelector />
 <div style="border: 1px solid #ccc; margin: 2px">
-    <DxDashboard style="width: 100%; height: 800px;" Endpoint="api/dashboard">
+    <DxDashboard style="width: 100%; height: 700px;" Endpoint="api/dashboard">
         <DxJSCustomization Identifier="dashboardEvents" />
     </DxDashboard>
 </div>
