@@ -1,27 +1,26 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace BlazorDashboardApp.Pages {
+namespace BlazorDashboardApp.Pages
+{
     [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
     [IgnoreAntiforgeryToken]
-    public class ErrorModel : PageModel {
-        public string RequestId { get; set; }
+    public class ErrorModel : PageModel
+    {
+        public string? RequestId { get; set; }
 
         public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
 
         private readonly ILogger<ErrorModel> _logger;
 
-        public ErrorModel(ILogger<ErrorModel> logger) {
+        public ErrorModel(ILogger<ErrorModel> logger)
+        {
             _logger = logger;
         }
 
-        public void OnGet() {
+        public void OnGet()
+        {
             RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier;
         }
     }
